Version 2.0
Application now able to see if HUD is already enabled so it won't by default disable the HUD when launching the application. This fix results in the application no longer disabling the HUD when quitting the application; this may be brought back as an optional feature in a future update.